+++ packages/server/src/sdk/workspace/ai/rag/index.ts
@@ -1,4 +1,4 @@
 export * from "./files"
 export * from "./sharepoint"
-export * as queue from "./queue"
+export * as ragQueue from "./ragQueue"
 export * as knowledgeSourceSyncQueue from "./knowledgeSourceSyncQueue"
