+++ packages/server/src/startup/index.ts
@@ -172,7 +172,7 @@ export async function startup(
       pro.sdk.licensing.client.getLicenseKey
     )
   )
-  queuePromises.push(rag.queue.init())
+  queuePromises.push(rag.ragQueue.init())
   queuePromises.push(rag.knowledgeSourceSyncQueue.init())
   queuePromises.push(
     rag.knowledgeSourceSyncQueue.rehydrateScheduledJobs().catch(err => {
